Fixes for double click bug on Anki 2.0
On Anki 2.0 double clicks make the addon behave as if the primary mouse button is constantly held down. The only way to return to normal behavior is to click the X button on the touchscreen menu. This fixes that issue.
I tested it on both Anki 2.1 and 2.0, so there should be no regressions.